% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_files_liwc.R
\name{clean_files_liwc}
\alias{clean_files_liwc}
\title{A .txt creator for LIWC analysis}
\usage{
clean_files_liwc(dataframe)
}
\arguments{
\item{dataframe}{A dataframe object.}
}
\description{
This function creates individual .txt files from participants cleaned tweets from a dataframe
}
\examples{
clean_files_liwc(dataframe)
}
\keyword{creation}
\keyword{text-file}
