% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mtest.R
\name{cor_mtest}
\alias{cor_mtest}
\title{Correlation significance}
\usage{
cor_mtest(mat, conf.level = 0.95)
}
\arguments{
\item{mat}{A correlation matrix.}

\item{conf.level}{The confidence level.}
}
\description{
This function tests the significance of correlations
}
\examples{
cor_mtest(mat, conf.level = 0.95)
}
\keyword{correlation}
\keyword{significance}
