% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_density.R
\name{emoji_density}
\alias{emoji_density}
\title{Emoji density function}
\usage{
emoji_density(dataframe)
}
\arguments{
\item{dataframe}{A data frame.}
}
\description{
Emoji density function
}
\examples{
emoji_density(dataframe)
}
\keyword{Emoji}
\keyword{function}
\keyword{test}
