% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_density.R
\name{emoji_density_test}
\alias{emoji_density_test}
\title{Test function}
\usage{
emoji_density_test(dataframe)
}
\arguments{
\item{dataframe}{A data frame.}
}
\description{
Emoji test function
}
\examples{
emoji_density_test(dataframe)

}
\keyword{Emoji}
\keyword{function}
\keyword{test}
