% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.gam.R
\name{ggplot.gam}
\alias{ggplot.gam}
\title{GAM plotting}
\usage{
ggplot.gam(model, type = "conditional", res = FALSE)
}
\arguments{
\item{model}{GAM object}

\item{type}{"conditional" or ?}

\item{res}{Residuals TRUE/FALSE}
}
\description{
Emoji test function
}
\examples{
ggplot.gam(model, type="conditional", res=FALSE)
}
\keyword{GAM}
\keyword{plotting}
