% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_omegas_anova.R
\name{partial_omegas}
\alias{partial_omegas}
\title{Partial omegas for ANOVAs}
\usage{
partial_omegas(mod)
}
\arguments{
\item{mod}{ANOVA model object}
}
\description{
Partial omegas
}
\examples{
partial_omegas(mod)
}
\keyword{ANOVA}
\keyword{omegas}
\keyword{partial}
