% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_processing.R
\name{text_processing}
\alias{text_processing}
\title{NTN}
\usage{
text_processing(x)
}
\arguments{
\item{x}{character vector}
}
\description{
This function is used to normalize noisy (GERMAN) tweets. use with lapply on character vector in dataframe
}
\examples{
text_processing(x)
}
\keyword{Text}
\keyword{normalization}
