% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_processing.R
\name{text_processing_with_punct}
\alias{text_processing_with_punct}
\title{NTN with punct}
\usage{
text_processing_with_punct(x)
}
\arguments{
\item{x}{character vector}
}
\description{
This function is used to normalize noisy (GERMAN) tweets. use with lapply on character vector in dataframe
}
\examples{
text_processing_with_punct(x)
}
\keyword{Text}
\keyword{normalization}
