% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_100.R
\name{time_to_25_part}
\alias{time_to_25_part}
\title{Time to 25 tweets}
\usage{
time_to_25_part(dataframe)
}
\arguments{
\item{dataframe}{A data frame object}
}
\description{
This function allows you to calculate the amount of time it takes a part to write 25 tweets (in days).
It is based on the data returned from a twitteR object and individual numerical participant ids genderated
for each screenName.
}
\examples{
time_to_25_part(dataframe)
}
\keyword{25}
\keyword{Time}
\keyword{to}
